+++ history.rst
@@ -4,6 +4,17 @@ Google App Engine - Address Book - Data Backend
 
 
 =============
+Version 0.1.4
+=============
+
+2015-09-15
+
+- New model for address changes: AddressHistory
+
+- *get_addresses* now accept the parameter *order_by* for sorting the result.
+
+
+=============
 Version 0.1.3
 =============